+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # pyspellchecker
 
+## Future Release
+* Added class method to be able to get a listing of all supported languages
+
 ## Version 0.6.2
 * Add ability to load multiple languages at once; [see discussion](https://github.com/barrust/pyspellchecker/discussions/97)
 * Fix default tokenizer to not enforce lower case; [#99](https://github.com/barrust/pyspellchecker/issues/99)